apps/fl_server/pyproject.toml
# Poetry
[tool.poetry]
name = "synthema-fl-server"
description = "Application containing the component fl-server"
authors = [
    "Borja Arroyo Galende <borja.arroyog@upm.es>"
]
readme = "README.md"
# package-mode = false
version = "0.0.0"
packages = [
    { include = "fl_server", from = "src"}
]

[tool.poetry.dependencies]
python = "^3.11.0"
flwr = "1.9"
pandas = "^2.2"
mlflow-skinny = "^2.11"
numpy = "^1"
pydantic = "^2.8.2"
setuptools = "^72.1.0"
pika = "^1.3.2"
sqlmodel = "^0.0.21"
psutil = "^6.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2"
pytest-env = "^1.1"
pytest-cov = "^5.0"
ruff = "^0.5"
mypy = "^1.10"
pip-licenses = "^5.0.0"

[tool.mypy]
python_version = "3.11"

[tool.poetry-dynamic-versioning]
enable = true

# Test
[tool.pytest.ini_options]
addopts = [
    "--import-mode=importlib",
]

[tool.pytest_env]
MLFLOW_URL = { value = "http://localhost:30246", skip_if_set = true }
NODE_NAME = { value = "test", skip_if_set = true }
DATA_PATH = { value = "common/data/iris.csv", skip_if_set = true }

[tool.pytest_coverage.run]
omit = ["tests/*"]

[tool.ruff]
target-version = "py311"

# Build system
[build-system]
requires = ["poetry-core>=1.0.0", "poetry-dynamic-versioning>=1.0.0,<2.0.0"]
build-backend = "poetry_dynamic_versioning.backend"
